+++ packages/relay-kit/src/index.ts
@@ -2,8 +2,7 @@ export * from './packs/gelato/GelatoRelayPack'
 export * from './packs/gelato/types'
 
 export * from './packs/safe-4337/Safe4337Pack'
-export * from './packs/safe-4337/estimators/pimlico'
-export * from './packs/safe-4337/estimators/alchemy'
+export * from './packs/safe-4337/estimators'
 export * from './packs/safe-4337/types'
 
 export * from './RelayKitBasePack'
