+++ packages/relay-kit/src/packs/safe-4337/estimators/PimlicoFeeEstimator.ts
@@ -1,8 +1,8 @@
 import { ethers } from 'ethers'
-import { EstimateFeeFunctionParams, EstimateGasData, IFeeEstimator } from '../types'
+import { EstimateFeeFunctionProps, EstimateGasData, IFeeEstimator } from '../types'
 
 export class PimlicoFeeEstimator implements IFeeEstimator {
-  async setupEstimation({ bundlerUrl }: EstimateFeeFunctionParams): Promise<EstimateGasData> {
+  async setupEstimation({ bundlerUrl }: EstimateFeeFunctionProps): Promise<EstimateGasData> {
     const bundlerClient = new ethers.JsonRpcProvider(bundlerUrl, undefined, {
       batchMaxCount: 1
     })
@@ -12,7 +12,7 @@ export class PimlicoFeeEstimator implements IFeeEstimator {
     return feeData
   }
 
-  async adjustEstimation({ userOperation }: EstimateFeeFunctionParams): Promise<EstimateGasData> {
+  async adjustEstimation({ userOperation }: EstimateFeeFunctionProps): Promise<EstimateGasData> {
     return {
       verificationGasLimit:
         userOperation.verificationGasLimit + userOperation.verificationGasLimit / 2n
