+++ packages/relay-kit/src/packs/safe-4337/estimators/index.ts
@@ -1,0 +1,7 @@
+import { PimlicoFeeEstimator } from './PimlicoFeeEstimator'
+import { AlchemyFeeEstimator } from './AlchemyFeeEstimator'
+
+export const estimators = {
+  PimlicoFeeEstimator,
+  AlchemyFeeEstimator
+}
